{"schema":"decodex/commit/1","summary":"Modularize API MCP and domain tests","authority":"manual"}
