+++ packages/elf-domain/src/writegate/tests.rs
@@ -1,293 +1,4 @@
-use serde_json::Map;
+pub(crate) mod config;
 
-use crate::writegate::{
-	self, NoteInput, RejectCode, WritePolicy, WritePolicyResult, WriteRedaction,
-	WriteRedactionResult,
-};
-use elf_config::{
-	Chunking, Config, EmbeddingProviderConfig, Lifecycle, LlmProviderConfig, Memory, MemoryPolicy,
-	Postgres, ProviderConfig, Providers, Qdrant, Ranking, RankingBlend, RankingBlendSegment,
-	RankingDeterministic, RankingDeterministicDecay, RankingDeterministicHits,
-	RankingDeterministicLexical, RankingDiversity, RankingRetrievalSources, ReadProfiles,
-	ScopePrecedence, ScopeWriteAllowed, Scopes, Search, SearchCache, SearchDynamic,
-	SearchExpansion, SearchExplain, SearchGraphContext, SearchPrefilter, SearchRecursive, Security,
-	Service, Storage, TtlDays,
-};
-
-fn test_ranking() -> Ranking {
-	Ranking {
-		recency_tau_days: 60.0,
-		tie_breaker_weight: 0.1,
-		deterministic: RankingDeterministic {
-			enabled: false,
-			lexical: RankingDeterministicLexical {
-				enabled: false,
-				weight: 0.05,
-				min_ratio: 0.3,
-				max_query_terms: 16,
-				max_text_terms: 1_024,
-			},
-			hits: RankingDeterministicHits {
-				enabled: false,
-				weight: 0.05,
-				half_saturation: 8.0,
-				last_hit_tau_days: 14.0,
-			},
-			decay: RankingDeterministicDecay { enabled: false, weight: 0.05, tau_days: 30.0 },
-		},
-		blend: RankingBlend {
-			enabled: true,
-			rerank_normalization: "rank".to_string(),
-			retrieval_normalization: "rank".to_string(),
-			segments: vec![
-				RankingBlendSegment { max_retrieval_rank: 3, retrieval_weight: 0.8 },
-				RankingBlendSegment { max_retrieval_rank: 10, retrieval_weight: 0.5 },
-				RankingBlendSegment { max_retrieval_rank: 1_000_000, retrieval_weight: 0.2 },
-			],
-		},
-		diversity: RankingDiversity {
-			enabled: true,
-			sim_threshold: 0.88,
-			mmr_lambda: 0.7,
-			max_skips: 64,
-		},
-		retrieval_sources: RankingRetrievalSources {
-			fusion_weight: 1.0,
-			structured_field_weight: 1.0,
-			fusion_priority: 1,
-			structured_field_priority: 0,
-		},
-	}
-}
-
-fn config() -> Config {
-	Config {
-		service: Service {
-			http_bind: "127.0.0.1:8080".to_string(),
-			mcp_bind: "127.0.0.1:8082".to_string(),
-			admin_bind: "127.0.0.1:8081".to_string(),
-			log_level: "info".to_string(),
-		},
-		storage: Storage {
-			postgres: Postgres {
-				dsn: "postgres://user:pass@localhost/db".to_string(),
-				pool_max_conns: 1,
-			},
-			qdrant: Qdrant {
-				url: "http://localhost".to_string(),
-				collection: "mem_notes_v2".to_string(),
-				docs_collection: "doc_chunks_v1".to_string(),
-				vector_dim: 4_096,
-			},
-		},
-		providers: Providers {
-			embedding: dummy_embedding_provider(),
-			rerank: dummy_provider(),
-			llm_extractor: dummy_llm_provider(),
-		},
-		scopes: Scopes {
-			allowed: vec!["agent_private".to_string()],
-			read_profiles: ReadProfiles {
-				private_only: vec!["agent_private".to_string()],
-				private_plus_project: vec!["agent_private".to_string()],
-				all_scopes: vec!["agent_private".to_string()],
-			},
-			precedence: ScopePrecedence { agent_private: 30, project_shared: 20, org_shared: 10 },
-			write_allowed: ScopeWriteAllowed {
-				agent_private: true,
-				project_shared: true,
-				org_shared: true,
-			},
-		},
-		memory: Memory {
-			max_notes_per_add_event: 3,
-			max_note_chars: 10,
-			dup_sim_threshold: 0.9,
-			update_sim_threshold: 0.8,
-			candidate_k: 10,
-			top_k: 5,
-			policy: MemoryPolicy { rules: vec![] },
-		},
-		search: Search {
-			expansion: SearchExpansion {
-				mode: "off".to_string(),
-				max_queries: 4,
-				include_original: true,
-			},
-			dynamic: SearchDynamic { min_candidates: 10, min_top_score: 0.12 },
-			prefilter: SearchPrefilter { max_candidates: 0 },
-			cache: SearchCache {
-				enabled: true,
-				expansion_ttl_days: 7,
-				rerank_ttl_days: 7,
-				max_payload_bytes: Some(262_144),
-			},
-			explain: SearchExplain {
-				retention_days: 7,
-				capture_candidates: false,
-				candidate_retention_days: 2,
-				write_mode: "outbox".to_string(),
-			},
-			recursive: SearchRecursive {
-				enabled: false,
-				max_depth: 2,
-				max_children_per_node: 4,
-				max_nodes_per_scope: 32,
-				max_total_nodes: 256,
-			},
-			graph_context: SearchGraphContext {
-				enabled: false,
-				max_facts_per_item: 16,
-				max_evidence_notes_per_fact: 16,
-			},
-		},
-		ranking: test_ranking(),
-		lifecycle: Lifecycle {
-			ttl_days: TtlDays {
-				plan: 1,
-				fact: 2,
-				preference: 0,
-				constraint: 0,
-				decision: 0,
-				profile: 0,
-			},
-			purge_deleted_after_days: 30,
-			purge_deprecated_after_days: 180,
-		},
-		security: Security {
-			bind_localhost_only: true,
-			reject_non_english: true,
-			redact_secrets_on_write: true,
-			evidence_min_quotes: 1,
-			evidence_max_quotes: 2,
-			evidence_max_quote_chars: 320,
-			auth_mode: "off".to_string(),
-			auth_keys: vec![],
-		},
-		chunking: Chunking {
-			enabled: true,
-			max_tokens: 512,
-			overlap_tokens: 128,
-			tokenizer_repo: "REPLACE_ME".to_string(),
-		},
-		context: None,
-		mcp: None,
-	}
-}
-
-fn dummy_embedding_provider() -> EmbeddingProviderConfig {
-	EmbeddingProviderConfig {
-		provider_id: "p".to_string(),
-		api_base: "http://localhost".to_string(),
-		api_key: "key".to_string(),
-		path: "/".to_string(),
-		model: "m".to_string(),
-		dimensions: 3,
-		timeout_ms: 1_000,
-		default_headers: Map::new(),
-	}
-}
-
-fn dummy_provider() -> ProviderConfig {
-	ProviderConfig {
-		provider_id: "p".to_string(),
-		api_base: "http://localhost".to_string(),
-		api_key: "key".to_string(),
-		path: "/".to_string(),
-		model: "m".to_string(),
-		timeout_ms: 1_000,
-		default_headers: Map::new(),
-	}
-}
-
-fn dummy_llm_provider() -> LlmProviderConfig {
-	LlmProviderConfig {
-		provider_id: "p".to_string(),
-		api_base: "http://localhost".to_string(),
-		api_key: "key".to_string(),
-		path: "/".to_string(),
-		model: "m".to_string(),
-		temperature: 0.1,
-		timeout_ms: 1_000,
-		default_headers: Map::new(),
-	}
-}
-
-#[test]
-fn rejects_long_text() {
-	let cfg = config();
-	let note = NoteInput {
-		note_type: "fact".to_string(),
-		scope: "agent_private".to_string(),
-		text: "12345678901".to_string(),
-	};
-
-	assert_eq!(writegate::writegate(&note, &cfg), Err(RejectCode::RejectTooLong));
-}
-
-#[test]
-fn rejects_invalid_type() {
-	let cfg = config();
-	let note = NoteInput {
-		note_type: "other".to_string(),
-		scope: "agent_private".to_string(),
-		text: "hello".to_string(),
-	};
-
-	assert_eq!(writegate::writegate(&note, &cfg), Err(RejectCode::RejectInvalidType));
-}
-
-#[test]
-fn detects_secret_patterns() {
-	assert!(writegate::contains_secrets("password: hunter2"));
-}
-
-#[test]
-fn applies_empty_policy_as_noop() {
-	let policy = WritePolicy::default();
-
-	assert_eq!(
-		writegate::apply_write_policy("keep this", Some(&policy)),
-		Ok(WritePolicyResult {
-			transformed: "keep this".to_string(),
-			..WritePolicyResult::default()
-		})
-	);
-}
-
-#[test]
-fn applies_exclusion_span() {
-	let policy = WritePolicy {
-		exclusions: vec![crate::writegate::WriteSpan { start: 4, end: 9 }],
-		redactions: vec![],
-	};
-	let actual = writegate::apply_write_policy("hello world", Some(&policy))
-		.expect("policy apply should succeed");
-
-	assert_eq!(actual.transformed, "hellld");
-	assert_eq!(actual.audit.exclusions, vec![crate::writegate::WriteSpan { start: 4, end: 9 }]);
-	assert!(actual.audit.redactions.is_empty());
-}
-
-#[test]
-fn applies_simple_replacement_redaction() {
-	let policy = WritePolicy {
-		exclusions: vec![],
-		redactions: vec![WriteRedaction::Replace {
-			span: crate::writegate::WriteSpan { start: 4, end: 5 },
-			replacement: "***".to_string(),
-		}],
-	};
-	let actual = writegate::apply_write_policy("secret", Some(&policy))
-		.expect("policy apply should succeed");
-
-	assert_eq!(actual.transformed, "secr***t");
-	assert_eq!(
-		actual.audit.redactions,
-		vec![WriteRedactionResult {
-			span: crate::writegate::WriteSpan { start: 4, end: 5 },
-			replacement: "***".to_string(),
-		}]
-	);
-	assert!(actual.audit.exclusions.is_empty());
-}
+mod policy;
+mod validation;
